requirements.txt
absl-py==2.1.0
aioice==0.9.0
aiortc==1.9.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
av==12.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.0
cvzone==1.6.1
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
flatbuffers==24.12.23
fonttools==4.55.3
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
jax==0.4.38
jaxlib==0.4.38
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mediapipe==0.10.20
ml_dtypes==0.5.1
narwhals==1.21.1
numpy==1.26.4
openai==1.59.5
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==4.25.5
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
pyee==12.1.1
Pygments==2.19.1
pylibsrtp==0.10.0
pyOpenSSL==24.3.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scipy==1.15.0
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
streamlit==1.41.1
streamlit-webrtc==0.47.9
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
